+++ src/Schema/Postgresql/Table.php
@@ -43,7 +43,7 @@ class Table extends AbstractTable
         // pg_get_expr on adbin, not compatible with pgsql < 9
         $adColName = ($version < 12 ? 'd.adsrc' : 'pg_get_expr(d.adbin,d.adrelid) AS adsrc');
 
-        $sql = "SELECT a.attname, a.attnotnull, a.atthasdef, a.attlen, a.atttypmod,
+        $sql = "SELECT a.attname, a.attnotnull, a.atthasdef, a.attlen, a.atttypmod, a.attgenerated,
                 FORMAT_TYPE(a.atttypid, a.atttypmod) AS type, a.attndims,
                 $adColName, co.contype AS primary, co.conname
             FROM pg_attribute AS a
@@ -62,11 +62,12 @@ class Table extends AbstractTable
             $notNull = ($line->attnotnull == 't');
             $default = $line->adsrc;
             $hasDefault = ($line->atthasdef == 't');
+            $generated = ($line->attgenerated != '');
             if ($type == 'boolean' && $hasDefault) {
                 $default = (strtolower($default) === 'true');
             }
 
-            $col = new Column($name, $type, $length, $hasDefault, $default, $notNull);
+            $col = new Column($name, $type, $length, $hasDefault, $default, $notNull, $generated);
 
             $typeinfo = $tools->getTypeInfo($type);
             if (is_string($default) && preg_match('/^nextval\(([^\)]*)\)$/', $default, $m)) {