+++ tests/docker-conf/php/resetdb.php
@@ -15,6 +15,7 @@ while($tryAgain) {
     pg_query($cnx, "drop table if exists product_test");
     pg_query($cnx, "drop table if exists labels_test");
     pg_query($cnx, "drop table if exists item_array_text");
+    pg_query($cnx, "drop table if exists generated_column_test");
 
     pg_query($cnx, "CREATE TABLE product_test (
         id serial NOT NULL,
@@ -51,6 +52,15 @@ while($tryAgain) {
     myintegers3 integer[3][2] NOT NULL
 )");
 
+
+    pg_query($cnx, "CREATE TABLE generated_column_test (
+    id serial NOT NULL,
+    description character varying(150) NOT NULL,
+    amount integer,
+    change numeric,
+    total numeric GENERATED ALWAYS AS (amount * change) STORED
+)");
+
     pg_query($cnx, "SELECT pg_catalog.setval(pg_catalog.pg_get_serial_sequence('products', 'id'), 1, false)");
 
     pg_query($cnx, "ALTER TABLE ONLY labels_test ADD CONSTRAINT labels_test_pkey PRIMARY KEY (\"key\", lang)");
@@ -64,7 +74,11 @@ while($tryAgain) {
 
     pg_query($cnx, "ALTER TABLE ONLY products ADD CONSTRAINT products_pkey PRIMARY KEY (id)");
 
+
     pg_query($cnx, "ALTER TABLE ONLY item_array_text ADD CONSTRAINT item_array_text_pkey PRIMARY KEY (id)");
+
+    pg_query($cnx, "SELECT pg_catalog.setval(pg_catalog.pg_get_serial_sequence('generated_column_test', 'id'), 1, false)");
+    pg_query($cnx, "ALTER TABLE ONLY generated_column_test ADD CONSTRAINT generated_column_test_pkey PRIMARY KEY (id)");
 
     pg_close($cnx);
 }